BWS-PKG - Bug 30541 - Resolve return claim works but "hangs" if MarkLostItemsAsReturned is set for return claims
Basically, if MarkLostItemsAsReturned is set for return claims, the code will crash while trying to pull the checkout so it can update the lost status of the item. This is because the code assumes there must be a checkout and never checks the old_checkouts table.

This bug appears to only exist in 21.05 at this time. It was fixed by a combination of bugs 29495 and 28588.

Test Plan:
1) Set MarkLostItemsAsReturned for return claims
2) Mark a checkout as claims returned
3) Verify the checkout is no longer on the patron record
4) Resolve the claim
5) Note the attempt "hangs" but refreshing the page shows it worked
6) Apply this patch, restart all the things!
7) Repeat steps 2-4
8) It works!