BWS-PKG - Bug 38912: Add option to include '_doc' type in ES record view
This bug has been filed several times and seems to relate to the ES version,
the version of Search::Elasticsearch and maybe other factors.

Because the version of ES being used may vary, and such the necessary perl module version
needed we should add an easy toggle to allow this feature to work, including the type if needed
and not if not.

To test:
 1 - Apply patches, updatedatabase
 2 - Enable Elasticsearch in Koha
 3 - View a record details page
 4 - Click 'Elastic record: Show'
 5 - You should see the record
 6 - Toggle ElasticsearchIncludeDocType system preference
 7 - You get a 404
 8 - Toggle it back
 9 - Record shows
10 - sudo apt-get install libsearch-elasticsearch-perl=6.81-1-koha1
11 - Restart all
12 - View ES record - it fails!
13 - Toggle the syspref
14 - It works!
15 - Sign off